requirements.txt
selenium
fake_useragent
pandas
beautifulsoup4
tqdm
requests
sqlalchemy
psycopg2-binary
fastapi
uvicorn
sqlalchemy
pydantic
pyodbc
python-dotenv
schedule
python-jose[cryptography]
python-multipart
python-docx